requirements.txt
Django==1.9
djangorestframework==3.4.6
Pillow==3.1.0
wheel==0.24.0
